+++ lib/legion/transport/messages/dynamic.rb
@@ -14,7 +14,9 @@ module Legion::Transport::Messages # rubocop:disable Style/ClassAndModuleChildre
     end
 
     def routing_key
-      "#{function.runner.extension.values[:name]}.#{function.runner.values[:name]}.#{function.values[:name]}"
+      key = "#{function.runner.extension.values[:name]}.#{function.runner.values[:name]}.#{function.values[:name]}"
+      Legion::Logging.debug "Dynamic routing_key=#{key} function_id=#{@options[:function_id]}" if defined?(Legion::Logging)
+      key
     end
 
     def exchange